+++ ocdsextensionregistry/util.py
@@ -5,7 +5,7 @@ from urllib.parse import urlsplit
 from zipfile import ZipFile
 
 import requests
-import requests_cache
+from requests_cache import CachedSession
 
 from .exceptions import UnknownLatestVersion
 
@@ -17,6 +17,8 @@ else:
     file_uri_offset = 7
 
 default_minor_version = '1.1'
+
+session = CachedSession(backend='memory')
 
 
 def json_dump(data, io):
@@ -56,8 +58,7 @@ def _resolve(data_or_url):
             with open(data_or_url[file_uri_offset:]) as f:
                 return f.read()
 
-        with requests_cache.enabled(backend='memory'):
-            response = requests.get(data_or_url)
+        response = session.get(data_or_url)
         response.raise_for_status()
         return response.text
 
@@ -78,8 +79,7 @@ def _resolve_zip(url, root=''):
                 for file in sorted(files):
                     zipfile.write(os.path.join(root, file), arcname=f'zip/{file}')
     else:
-        with requests_cache.enabled(backend='memory'):
-            response = requests.get(url, allow_redirects=True)
+        response = session.get(url, allow_redirects=True)
         response.raise_for_status()
         io = BytesIO(response.content)